+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
   is the official name for the DNAx FUSE filesystem.
 - [Prefer](https://github.com/dnanexus/dxWDL/issues/309) v2 instances over v1 instances.
 - Improving marshalling of WDL values to JSON.
+- Allow a WDL input that is a map where the key is a file.
 
 ## 1.30
 - Using dxfs2 to stream files. This replaces `dx cat`, which was the previous solution.